Create README
listOfDevices&aColorPicker This is my first project. It visualizes a list of devices with ON/OFF buttons. Once a button is clicked, its style and title are changed. Clicking on the device row (outside the button) user is transferred to page 2, "Device settings". There, the user will be able to choose a color from a color wheel palette. Upon a choice the device sends an AJAX request submitting the selected color and delivers a result in JSON format.